pyproject.toml
[project]
name = "open-mirroring-faker"
version = "0.1.0"
description = "Generate synthetic data for Microsoft Fabric Open Mirroring demos"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.12"
authors = [{ name = "Luke Moloney" }]
dependencies = [
    "azure-storage-file-datalake>=12.14",
    "azure-identity>=1.15",
    "pyarrow>=15.0",
    "click>=8.0",
    "pyyaml>=6.0",
    "faker>=30.0",
    "python-dotenv>=1.0",
]

[project.scripts]
omf = "open_mirroring_faker.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/open_mirroring_faker"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "ruff>=0.4",
    "pyright>=1.1",
]

[tool.ruff]
target-version = "py312"
line-length = 120
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM", "RUF"]

[tool.pyright]
pythonVersion = "3.12"
typeCheckingMode = "basic"
venvPath = "."
venv = ".venv"

[tool.pytest.ini_options]
testpaths = ["tests"]
